requirements.txt
langchain
langchain-community
faiss-cpu
euriai
langchain-huggingface
pypdf
python-dotenv
sentence-transformers
euriai
streamlit
